feat: split off floor store from printer store for a cleaner structure
