+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@fdm-monster/client",
-  "version": "1.1.5",
+  "version": "1.2.0",
   "private": false,
   "author": "David Zwart",
   "license": "AGPL-3.0-or-later",
