+++ src/components/PrinterGrid/HomeToolbar.vue
@@ -14,12 +14,12 @@
       </v-btn>
     </v-btn-toggle>
 
-    <v-btn color="primary" class="mt-0 ml-6" v-if="!printersStore.printers?.length" to="/printers">
+    <v-btn color="primary" class="mt-0 ml-6" v-if="!printerStore.printers?.length" to="/printers">
       You have no printers. Click here to start!
     </v-btn>
-    <v-alert color="primary" class="mt-4 ml-12" v-if="printersStore.floorlessPrinters.length">
+    <v-alert color="primary" class="mt-4 ml-12" v-if="floorStore.floorlessPrinters.length">
       <v-icon>warning</v-icon>
-      {{ printersStore.floorlessPrinters.length }} unplaced printer(s)!</v-alert
+      {{ floorStore.floorlessPrinters.length }} unplaced printer(s)!</v-alert
     >
     <div class="ma-4 pt-6">
       <v-switch v-model="gridStore.gridEditMode" label="Printer Relocate Mode"></v-switch>
@@ -30,15 +30,17 @@
 
 <script lang="ts">
 import { defineComponent } from "vue";
-import { usePrintersStore } from "@/store/printers.store";
+import { usePrinterStore } from "../../store/printer.store";
 import { useGridStore } from "../../store/grid.store";
+import { useFloorStore } from "../../store/floor.store";
 
 export default defineComponent({
   name: "HomeToolbar",
   components: {},
   setup() {
     return {
-      printersStore: usePrintersStore(),
+      printerStore: usePrinterStore(),
+      floorStore: useFloorStore(),
       gridStore: useGridStore(),
     };
   },
@@ -51,12 +53,12 @@ export default defineComponent({
   },
   computed: {
     floors() {
-      return this.printersStore.floors;
+      return this.floorStore.floors;
     },
   },
   methods: {
     changeFloorIndex(index: any) {
-      this.printersStore.changeSelectedFloorByIndex(index);
+      this.floorStore.changeSelectedFloorByIndex(index);
       this.selectedFloorToggleIndex = index;
     },
   },
