+++ src/components/Settings/SettingsView.vue
@@ -31,7 +31,7 @@
 
 <script lang="ts">
 import { defineComponent } from "vue";
-import { usePrintersStore } from "../../store/printers.store";
+import { usePrinterStore } from "../../store/printer.store";
 
 interface Data {
   items: any[];
@@ -42,7 +42,7 @@ export default defineComponent({
   components: {},
   setup: () => {
     return {
-      printersStore: usePrintersStore(),
+      printersStore: usePrinterStore(),
     };
   },
   async created() {},
